Merge marlowe tutorial into main docs
This merges the Marlowe tutorials into the Sphinx doc site.

This has a few advantages:
- We can cross-reference between them more easily (there are a couple of
exmaples of this even in the docs to date), and share explanatory content.
- The Sphinx site makes it easier to structure the docs (as opposed to
with Asiidoctor where it was all one page).
- We can engage with documentation writers and infrastructure people
with a combined front.

I haven't touched the content and structure much, which could definitely do with work.

(I also optimized the PNG sizes with `optipng`, since otherwise it's
quite a lot of bytes even just moving them...)